+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Changed `ImagePosition` to string union with additional options
+- Converted `ImageFit` and `FlipDirection` types to string unions to match `ImagePosition`
+
 ## [0.3.18] - 2022-04-23
 
 ### CI